+++ packages/content-types/src/post-types/actions/quick-edit.tsx
@@ -23,13 +23,15 @@ import {
 	defaultForm,
 	hasArchiveField,
 	hierarchicalField,
-	pluralLabelField,
 	publicField,
-	singularLabelField,
-	statusField,
 	useSlugField,
 	useTaxonomiesField,
 } from '../fields';
+import {
+	pluralLabelField,
+	singularLabelField,
+	statusField,
+} from '../../utils/fields';
 import type { PostTypeFormData } from '../types';
 import { serializeForSave } from '../utils';
 import { POST_TYPE_ENTITY } from '../../constants';
@@ -47,17 +49,18 @@ function QuickEditPostTypeModal( {
 	const slugField = useSlugField( item.slug, data.slug );
 	const taxonomiesField = useTaxonomiesField();
 
-	const fields = useMemo< Field< PostTypeFormData >[] >(
-		() => [
-			pluralLabelField,
-			singularLabelField,
-			slugField,
-			taxonomiesField,
-			publicField,
-			hierarchicalField,
-			hasArchiveField,
-			statusField,
-		],
+	const fields = useMemo(
+		() =>
+			[
+				pluralLabelField,
+				singularLabelField,
+				slugField,
+				taxonomiesField,
+				publicField,
+				hierarchicalField,
+				hasArchiveField,
+				statusField,
+			] as Field< PostTypeFormData >[],
 		[ slugField, taxonomiesField ]
 	);
 
@@ -100,7 +103,12 @@ function QuickEditPostTypeModal( {
 	}
 
 	return (
-		<>
+		<form
+			onSubmit={ ( event ) => {
+				event.preventDefault();
+				onSave();
+			} }
+		>
 			<Stack
 				className="dataviews-action-modal__quick-edit-post-type-header"
 				direction="row"
@@ -146,15 +154,15 @@ function QuickEditPostTypeModal( {
 				<Button
 					__next40pxDefaultSize
 					variant="primary"
+					type="submit"
 					isBusy={ isSaving }
-					disabled={ isSaving }
+					disabled={ isSaving || ! isValid }
 					accessibleWhenDisabled
-					onClick={ onSave }
 				>
 					{ __( 'Done' ) }
 				</Button>
 			</Stack>
-		</>
+		</form>
 	);
 }