+++ packages/content-types/src/post-types/edit.tsx
@@ -37,20 +37,18 @@ import {
 	insertIntoItemField,
 	itemsListField,
 	itemsListNavigationField,
+	labelsActionsField,
 	labelsForm,
 	menuNameField,
 	newItemField,
 	notFoundField,
 	notFoundInTrashField,
 	parentItemColonField,
-	pluralLabelField,
 	publicField,
 	removeFeaturedImageField,
 	searchItemsField,
 	setFeaturedImageField,
 	showInRestField,
-	singularLabelField,
-	statusField,
 	supportsField,
 	uploadedToThisItemField,
 	useFeaturedImageField,
@@ -59,6 +57,11 @@ import {
 	viewItemField,
 	viewItemsField,
 } from './fields';
+import {
+	pluralLabelField,
+	singularLabelField,
+	statusField,
+} from '../utils/fields';
 import type { PostTypeFormData, PostTypeRecord } from './types';
 import { BLANK_RECORD, serializeForSave, toFormData } from './utils';
 import { NEW_ID, POST_TYPE_ENTITY, POST_TYPES_PATH } from '../constants';
@@ -135,45 +138,47 @@ function PostTypePage( {
 	const originalSlug = ! isAddMode ? initialData.slug : undefined;
 	const slugField = useSlugField( originalSlug, data.slug );
 	const taxonomiesField = useTaxonomiesField();
-	const fields = useMemo< Field< PostTypeFormData >[] >(
-		() => [
-			// General
-			pluralLabelField,
-			singularLabelField,
-			slugField,
-			descriptionField,
-			taxonomiesField,
-			supportsField,
-			publicField,
-			hierarchicalField,
-			hasArchiveField,
-			showInRestField,
-			statusField,
-			// Labels
-			menuNameField,
-			allItemsField,
-			addNewField,
-			addNewItemLabelField,
-			editItemField,
-			newItemField,
-			viewItemField,
-			viewItemsField,
-			searchItemsField,
-			notFoundField,
-			notFoundInTrashField,
-			parentItemColonField,
-			archivesField,
-			attributesField,
-			insertIntoItemField,
-			uploadedToThisItemField,
-			featuredImageField,
-			setFeaturedImageField,
-			removeFeaturedImageField,
-			useFeaturedImageField,
-			filterItemsListField,
-			itemsListNavigationField,
-			itemsListField,
-		],
+	const fields = useMemo(
+		() =>
+			[
+				// General
+				pluralLabelField,
+				singularLabelField,
+				slugField,
+				descriptionField,
+				taxonomiesField,
+				supportsField,
+				publicField,
+				hierarchicalField,
+				hasArchiveField,
+				showInRestField,
+				statusField,
+				// Labels
+				labelsActionsField,
+				menuNameField,
+				allItemsField,
+				addNewField,
+				addNewItemLabelField,
+				editItemField,
+				newItemField,
+				viewItemField,
+				viewItemsField,
+				searchItemsField,
+				notFoundField,
+				notFoundInTrashField,
+				parentItemColonField,
+				archivesField,
+				attributesField,
+				insertIntoItemField,
+				uploadedToThisItemField,
+				featuredImageField,
+				setFeaturedImageField,
+				removeFeaturedImageField,
+				useFeaturedImageField,
+				filterItemsListField,
+				itemsListNavigationField,
+				itemsListField,
+			] as Field< PostTypeFormData >[],
 		[ slugField, taxonomiesField ]
 	);
 
@@ -197,9 +202,6 @@ function PostTypePage( {
 				{
 					id: 'labels',
 					label: __( 'Labels' ),
-					description: __(
-						'Override the text WordPress shows in admin lists, menus, and forms. Leave blank to use defaults derived from the plural and singular names.'
-					),
 					layout: {
 						type: 'card',
 						isCollapsible: true,