+++ packages/grid/src/index.ts
@@ -2,6 +2,7 @@ export { DashboardGrid } from './grid';
 export type {
 	DashboardGridLayoutItem,
 	DashboardGridProps,
+	DragPreviewRenderProps,
 	ResizeDelta,
 	ResizeHandleRenderProps,
 } from './types';